Add ErrParamError
