+++ rest/httpx/requests_lt1.20.go
@@ -1,0 +1,76 @@
+//go:build !go1.20
+
+package httpx
+
+import (
+	"github.com/zeromicro/go-zero/core/mapping"
+	"github.com/zeromicro/go-zero/core/validation"
+	"github.com/zeromicro/go-zero/rest/internal/encoding"
+	"github.com/zeromicro/go-zero/rest/pathvar"
+	"io"
+	"net/http"
+)
+
+// Parse parses the request.
+func Parse(r *http.Request, v any) error {
+	if err := ParsePath(r, v); err != nil {
+		return err
+	}
+
+	if err := ParseForm(r, v); err != nil {
+		return err
+	}
+
+	if err := ParseHeaders(r, v); err != nil {
+		return err
+	}
+
+	if err := ParseJsonBody(r, v); err != nil {
+		return err
+	}
+
+	if valid, ok := v.(validation.Validator); ok {
+		return valid.Validate()
+	} else if val := validator.Load(); val != nil {
+		return val.(Validator).Validate(r, v)
+	}
+
+	return nil
+}
+
+// ParseHeaders parses the headers request.
+func ParseHeaders(r *http.Request, v any) error {
+	return encoding.ParseHeaders(r.Header, v)
+}
+
+// ParseForm parses the form request.
+func ParseForm(r *http.Request, v any) error {
+	params, err := GetFormValues(r)
+	if err != nil {
+		return err
+	}
+
+	return formUnmarshaler.Unmarshal(params, v)
+}
+
+// ParseJsonBody parses the post request which contains json in body.
+func ParseJsonBody(r *http.Request, v any) error {
+	if withJsonBody(r) {
+		reader := io.LimitReader(r.Body, maxBodyLen)
+		return mapping.UnmarshalJsonReader(reader, v)
+	}
+
+	return mapping.UnmarshalJsonMap(nil, v)
+}
+
+// ParsePath parses the symbols reside in url path.
+// Like http://localhost/bag/:name
+func ParsePath(r *http.Request, v any) error {
+	vars := pathvar.Vars(r)
+	m := make(map[string]any, len(vars))
+	for k, v := range vars {
+		m[k] = v
+	}
+
+	return pathUnmarshaler.Unmarshal(m, v)
+}
